test: modularize DINO AutoML coverage
